+++ Arex388.Carmine/Models/Event.cs
@@ -1,6 +1,4 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Arex388.Carmine;
+﻿namespace Arex388.Carmine;
 
 /// <summary>
 /// Event object.
@@ -9,24 +7,20 @@ public sealed class Event {
 	/// <summary>
 	/// Additional data about the event.
 	/// </summary>
-	[JsonPropertyName("event_data")]
 	public string? Data { get; init; }
 
 	/// <summary>
 	/// The event's end timestamp.
 	/// </summary>
-	[JsonPropertyName("end_time")]
 	public DateTime EndAt { get; init; }
 
 	/// <summary>
 	/// The event's type.
 	/// </summary>
-	[JsonPropertyName("event_type")]
 	public EventType Type { get; init; }
 	
 	/// <summary>
 	/// The event's start timestamp.
 	/// </summary>
-	[JsonPropertyName("start_time")]
 	public DateTime StartAt { get; init; }
 }