+++ Arex388.Carmine/Models/User.cs
@@ -1,6 +1,4 @@
-﻿using System.Text.Json.Serialization;
-
-namespace Arex388.Carmine;
+﻿namespace Arex388.Carmine;
 
 /// <summary>
 /// User object.
@@ -9,7 +7,6 @@ public sealed class User {
 	/// <summary>
 	/// Flag indicating if the user can share their vehicle's ETA.
 	/// </summary>
-	[JsonPropertyName("can_share_eta")]
 	public bool CanShareVehiclesEta { get; init; }
 
 	/// <summary>
@@ -20,7 +17,6 @@ public sealed class User {
 	/// <summary>
 	/// The user's created timestamp in UTC.
 	/// </summary>
-	[JsonPropertyName("created")]
 	public DateTime CreatedAtUtc { get; init; }
 
 	/// <summary>
@@ -36,19 +32,16 @@ public sealed class User {
 	/// <summary>
 	/// Flag indicating if the user's email has been validated.
 	/// </summary>
-	[JsonPropertyName("validated")]
 	public bool IsEmailValidated { get; init; }
 
 	/// <summary>
 	/// The user's last activity timestamp.
 	/// </summary>
-	[JsonPropertyName("last_activity")]
 	public DateTime? LastActivityAt { get; init; }
 
 	/// <summary>
 	/// The user's last trip id.
 	/// </summary>
-	[JsonPropertyName("last_trip")]
 	public TripId? LastTripId { get; init; }
 
 	/// <summary>
@@ -59,7 +52,6 @@ public sealed class User {
 	/// <summary>
 	/// The user's phone.
 	/// </summary>
-	[JsonPropertyName("sms")]
 	public long? Phone { get; init; }
 
 	/// <summary>
@@ -75,12 +67,10 @@ public sealed class User {
 	/// <summary>
 	/// The user's status.
 	/// </summary>
-	[JsonPropertyName("active")]
 	public UserStatus Status { get; init; }
 
 	/// <summary>
 	/// The user's current vehicle id.
 	/// </summary>
-	[JsonPropertyName("current_vehicle")]
 	public VehicleId? VehicleId { get; init; }
 }